#53 예매 정보 추가, 결제 정보 추가
결제 정보 추상화(카카오 페이, 무통장 입금, 신용카드)
결제 정보에 따른 결제 정보 추가
결제 정보 로직 추가
int -> long 타입으로 변경

+++ src/main/java/com/younghun/hibusgo/controller/BusTerminalController.java
@@ -108,7 +108,7 @@ public class BusTerminalController {
    */
   @LoginCheck(userLevel = UserLevel.ADMIN)
   @DeleteMapping("/{id}")
-  public ResponseEntity<?> deleteBusTerminal(@PathVariable int id) {
+  public ResponseEntity<?> deleteBusTerminal(@PathVariable long id) {
     boolean isExistsTerminal =  busTerminalService.existsById(id);
 
     if (!isExistsTerminal) {
@@ -129,7 +129,7 @@ public class BusTerminalController {
   @PatchMapping()
   public ResponseEntity<?> updateBusTerminal(@RequestBody @Valid BusTerminalDto busTerminalDto) {
     String name = busTerminalDto.getName();
-    int regionId = busTerminalDto.getRegionId();
+    long regionId = busTerminalDto.getRegionId();
 
     boolean isExistsRegion = regionService.existsById(regionId);
