+++ src/main/java/com/younghun/hibusgo/service/RegionService.java
@@ -16,11 +16,11 @@ public class RegionService {
 
   private final RegionMapper regionMapper;
 
-  public Region findById(int id) {
+  public Region findById(long id) {
     return regionMapper.findById(id);
   }
   
-  public boolean existsById(int id) {
+  public boolean existsById(long id) {
     return regionMapper.existsById(id);
   }
 
@@ -47,7 +47,7 @@ public class RegionService {
     regionMapper.addRegion(region);
   }
 
-  public void deleteRegion(int id) {
+  public void deleteRegion(long id) {
     regionMapper.deleteRegion(id);
   }
 }
